Memory access problem.

+++ module-gui/gui/core/Context.hpp
@@ -17,7 +17,7 @@ namespace gui {
 class Context {
 protected:
 	int16_t  x, y;
-	uint16_t w, h;
+	uint32_t w, h;
 	uint8_t* data;
 public:
 	Context();
